python_client/requirements.txt
requests==2.18.4
google-auth==1.6.1
google-api-python-client==1.7.7
google-cloud-bigquery==1.8.1
google-cloud-storage==1.13.2
google-cloud-iam==0.1.0